% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNested.R
\name{boxp}
\alias{boxp}
\title{boxp}
\usage{
boxp(box)
}
\arguments{
\item{box}{The box inside which the prior probability is finite}
}
\value{
Box prior function
}
\description{
Create a box prior function
}
\details{
The prior function returned takes position in parameter space as
an argument and returns log-prior probability. Value of -998 is
returned outside the box and 0 inside the box
}
\author{
bnikolic
}
