% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNested.R
\name{CPChain}
\alias{CPChain}
\title{CPChain}
\usage{
CPChain(s, proposer, n, llf, lpf, cs)
}
\arguments{
\item{s}{The starting point}

\item{proposer}{Function to propose new points}

\item{n}{Number of steps to make}

\item{llf}{Log-likelihood function}

\item{lpf}{Log-prior function}

\item{cs}{Live (or current) set}
}
\value{
Either the new point or if new point was not found FALSE
}
\description{
Constrained Prior Sampler using modified MCMC
}
\details{
Note that the performance of the overal nested sampler is close to
linearly dependent on the value of the parameter n. When  a good
proposer (in the sense that it proposes that are uncorrelated with
the starting or existing points but tend to satisfy the likelihood
constraint) for new points is used this value of n can be reduced
leading to better performance.
}
\author{
bnikolic
}
