% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsCommon.R
\name{Create_P}
\alias{Create_P}
\title{Create_P}
\usage{
Create_P(l, m, n, Rxy, y_zx_depenendent = FALSE)
}
\arguments{
\item{l}{|Z|}

\item{m}{|X|}

\item{n}{|Y|}

\item{Rxy}{Matrix of response function indicators}

\item{y_zx_dependent}{If set to true P will be calculated assuming Y is dependent on z and x, otherwise will assume y is only dependent on x.}
}
\description{
Creates a l\*m\*n x (m^l)\*(n^m) dimensional matrix if y_zx_dependent = FAlSE, creates l\*m\*n x (m^l)\*((n^m)+1) dimension matrix otherwise.
}
\keyword{internal}
