% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsCommon.R
\name{estimate_integral}
\alias{estimate_integral}
\title{estimate_integral}
\usage{
estimate_integral(N, S, d, llf, sample_theta)
}
\arguments{
\item{N}{Number of Repetitions}

\item{S}{Number of Starting Points}

\item{d}{number of dimension of point}

\item{llf}{log likelihood function}
}
\description{
Estimates multidimensional integral using nested sampling.
}
\keyword{internal}
