% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amenities.R
\name{get_amenities}
\alias{get_amenities}
\title{Retrieve the amenity types}
\usage{
get_amenities(limit = 50)
}
\arguments{
\item{limit}{Maximum number of results (Default:50)}
}
\value{
a dataframe of all amenities
}
\description{
Retrieve the amenity types
}
