% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amenities.R
\name{get_amenities_by_parksplace}
\alias{get_amenities_by_parksplace}
\title{Retrieve "places" within national parks that have different amenities}
\usage{
get_amenities_by_parksplace(park_code = NULL, limit = 50)
}
\arguments{
\item{park_code}{National park codes (e.g., 'ACAD')}

\item{limit}{Maximum number of results (Default: 50)}
}
\value{
Return a tibble that has the places within national parks that have different amenities
}
\description{
Retrieve "places" within national parks that have different amenities
}
