% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{get_articles}
\alias{get_articles}
\title{Retrieve articles created by national parks and other NPS entities}
\usage{
get_articles(park_code = NULL, state_code = NULL, limit = 50)
}
\arguments{
\item{park_code}{park codes}

\item{state_code}{state codes}

\item{limit}{Number of results to return per request. Default is 50.}
}
\value{
a tibble of articles
}
\description{
Retrieve articles created by national parks and other NPS entities
}
