% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campgrounds.R
\name{get_campgrounds}
\alias{get_campgrounds}
\title{Retrieve data about National Park Service campgrounds}
\usage{
get_campgrounds(park_code = NULL, state_code = NULL, limit = 50)
}
\arguments{
\item{park_code}{park codes}

\item{state_code}{state codes}

\item{limit}{Number of results to return per request. Default is 50.}
}
\value{
a tibble of articles
}
\description{
Retrieve data about National Park Service campgrounds
}
