% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks.R
\name{get_park_activities}
\alias{get_park_activities}
\title{Retrieve national parks that are related to particular categories of activities}
\usage{
get_park_activities(limit = 50)
}
\arguments{
\item{limit}{Maximum number of results (default: 50)}
}
\value{
A tibble of activities and national parks
}
\description{
Retrieve national parks that are related to particular categories of activities
}
