% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks.R
\name{get_parks}
\alias{get_parks}
\title{Retrieve National Park Service park data}
\usage{
get_parks(park_code = NULL, state_code = NULL, limit = 50)
}
\arguments{
\item{park_code}{Filter parks by park code (e.g., 'ACAD')}

\item{state_code}{Filter parks by state (e.g., 'CA')}

\item{limit}{Maximum number of results (default: 50)}
}
\value{
A tibble of park data
}
\description{
Retrieve National Park Service park data
}
