% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks.R
\name{get_passport_stamp_locations}
\alias{get_passport_stamp_locations}
\title{Retrieve location that have national park passport stamps}
\usage{
get_passport_stamp_locations(park_code = NULL, state_code = NULL, limit = 50)
}
\arguments{
\item{park_code}{Filter by park_code (e.g., 'CA')}

\item{state_code}{Filter by state (e.g., 'CA')}

\item{limit}{Maximum number of results (default: 50)}
}
\value{
A tibble of passport location data
}
\description{
Retrieve location that have national park passport stamps
}
