% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsInvalid.R
\name{M_air_excl}
\alias{M_air_excl}
\title{m_air}
\usage{
M_air_excl(Q, l, m, n)
}
\arguments{
\item{Q}{Histogram of dataset (l\*m\*n vector)}

\item{l}{|Z|}

\item{m}{|X|}

\item{n}{|Y|}
}
\value{
The probability that the observations where created from
a model which violates the as-if-randomness criterion but not the exclusion criterion
}
\description{
Calculates the marginal likelihood M_air_excel
}
