% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsInvalid.R
\name{M_excl}
\alias{M_excl}
\title{M_excl}
\usage{
M_excl(Q, l, m, n, N = sum(Q), S = sum(Q))
}
\arguments{
\item{Q}{Histogram of dataset (l\*m\*n vector)}

\item{l}{|Z|}

\item{m}{|X|}

\item{n}{|Y|}

\item{N}{Number of Repetitions for Nested Sampling}

\item{S}{Number of Starting Points for Nested Sampling}
}
\value{
The probability that the observations where created from a model which violates the exclusion criterion but not the as-if-randomness criterion
}
\description{
Calculates the marginal likelihood of M_excl
}
