% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsCommon.R
\name{make_proposer}
\alias{make_proposer}
\title{make_proposer}
\usage{
make_proposer(d, sample_theta)
}
\arguments{
\item{d}{number of dimensions of theta}

\item{N}{function to sample new theta}
}
\description{
Creates a function which proposes new random theta.
}
\keyword{internal}
