% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsCommon.R
\name{make_sampler}
\alias{make_sampler}
\title{make_sampler}
\usage{
make_sampler(d, s, sample_theta)
}
\arguments{
\item{d}{Dimensions of point}

\item{s}{Number of mcmc walker steps}
}
\value{
Sampler function which tries to find a point with a higher likelihood than the given point
}
\description{
Creates a sampler function which finds a new point with a higher log likelihood.
}
\keyword{internal}
