% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNested.R
\name{mkCovarianceSampler}
\alias{mkCovarianceSampler}
\title{mkCovarianceSampler}
\usage{
mkCovarianceSampler(s = 1)
}
\arguments{
\item{s}{Scale the covariances of the live set by this factor
before using them for sampling}
}
\value{
Constrained prior sampler
}
\description{
Use the covariance of the points in the live set to determine the
scales for sampling the prior space
}
\author{
bnikolic
}
