% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNested.R
\name{mkPriorSamplePred}
\alias{mkPriorSamplePred}
\title{mkPriorSamplePred}
\usage{
mkPriorSamplePred(s)
}
\arguments{
\item{s}{Row of the live set which is being replaced}
}
\value{
The predicate function
}
\description{
Create a function that will accept or reject proposals for the new sample
}
\author{
bnikolic
}
