% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNested.R
\name{nested.sample}
\alias{nested.sample}
\title{nested.sample}
\usage{
nested.sample(cs, llf, lpf, psampler, cout = rbind(), N = 1)
}
\arguments{
\item{cs}{The current/live/starting set of poings}

\item{llf}{The log-likelihood function (which should take one
argument which is the point in parameter space at which to evalue
the likelihood)}

\item{lpf}{the log-prior probability function (which should take
one argument, which is the point in parameter sample)}

\item{psampler}{The prior sampler to use to advance find sample a
new point  from the prior space}

\item{cout}{Optional parameter which is the output of a previous
run of the nested sampler. Allows easy restart of the algorithm}

\item{N}{Number of samples to make}
}
\value{
List of the live set  and the output of nested
sampling. See nested.summary and nested.hist2 on how to interpret
the output
}
\description{
Do nested sampling of the suplied pair of log-likelihood and prior
probability functions
}
\details{
This function is the top-level driver for the nested sampling
algorithm. It requires preparation of a starting set (using for
example sset.box) and a constrained prior space sampler (using for
example mkCovarianceSampler). It allows restart of the sampling if
inspection of outputs shows that the sampler may not have
converged
}
\author{
bnikolic
}
