% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNested.R
\name{nested.step}
\alias{nested.step}
\title{nested.step}
\usage{
nested.step(cs, llf, lpf, psampler)
}
\arguments{
\item{cs}{The current (live) set}

\item{llf}{Log-likelihood function}

\item{lpf}{Log-prior function}

\item{psampler}{The prior space sampler}
}
\value{
list (new current set, eliminated row)
}
\description{
Take one step of the nested sampler
}
\author{
bnikolic
}
