% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsMain.R
\name{nps.discretize}
\alias{nps.discretize}
\title{nps.discretize}
\usage{
nps.discretize(variables, d)
}
\arguments{
\item{variables}{Array of variables to be discretized}

\item{d}{number of classes (dimensions) to classify the given variables into}
}
\value{
Discretized array
}
\description{
Discretizes the given \code{variable} into \code{d} bins using Jenks natural breaks optimization
}
