% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsNecessary.R
\name{nps.necessary}
\alias{nps.necessary}
\title{testIc}
\usage{
nps.necessary(df)
}
\arguments{
\item{df}{Dataframe with z, x and y}
}
\description{
Tests the instrumental constraints on the given dataframe using entropy
}
