% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsMain.R
\name{nps.test}
\alias{nps.test}
\title{Main function of the package.}
\usage{
nps.test(df, l, m, n)
}
\arguments{
\item{df}{Dataframe with columns z,x and y}

\item{l}{Number of bins used to discretize Z}

\item{m}{Number of bins used to discretize X}

\item{n}{Number of bins used to discretize Y}
}
\value{
result object of the test whith the fields: nt, valid, invalid, ratio
}
\description{
Main function of the package.
}
