% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{nps_get}
\alias{nps_get}
\title{Internal function to make API requests}
\usage{
nps_get(endpoint, query = list(), api_key = Sys.getenv("NPS_API_KEY"))
}
\description{
Internal function to make API requests
}
\keyword{internal}
