% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsInvalid.R
\name{product_fraction}
\alias{product_fraction}
\title{Reduces out factors of fraction of products and calculates the fraction
Analog to prod(num)/prod(den)}
\usage{
product_fraction(num, den)
}
\arguments{
\item{num}{vector of factors of the numerator}

\item{den}{vector of factors of the denominator}
}
\description{
Reduces out factors of fraction of products and calculates the fraction
Analog to prod(num)/prod(den)
}
