% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNested.R
\name{randomEl}
\alias{randomEl}
\title{randomEl}
\usage{
randomEl(cs)
}
\arguments{
\item{cs}{The live (or current) set}
}
\value{
The selected element
}
\description{
Return a random element of the live set
}
\author{
bnikolic
}
