% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNested.R
\name{rectOffseter}
\alias{rectOffseter}
\title{rectOffseter}
\usage{
rectOffseter(scale)
}
\arguments{
\item{scale}{Array of scales (one for each parameter of the
problem) for generation of the offsets}
}
\value{
Function that generates the offset
}
\description{
Rectangular offseting with user supplied scales, pretty much the
simplest offsetter
}
\author{
bnikolic
}
