% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNested.R
\name{sset.box}
\alias{sset.box}
\title{sset.box}
\usage{
sset.box(box, nss, llfn)
}
\arguments{
\item{box}{Matrix with boundaries of the box}

\item{nss}{Number of elements of starting set}

\item{llfn}{Function to initilalise the log-likelihod}
}
\value{
The starting set as a frame
}
\description{
Create a starting set based on a box prior
}
\author{
bnikolic
}
