% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsCommon.R
\name{starting_points}
\alias{starting_points}
\title{starting_points}
\usage{
starting_points(N, d, sample_theta)
}
\arguments{
\item{N}{number of samples}

\item{d}{number of dimensions}
}
\description{
Creates a set of N starting Points for theta Z
}
\keyword{internal}
