% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsCommon.R
\name{XY_product}
\alias{XY_product}
\title{XY_product}
\usage{
XY_product(theta_xy, P, Q)
}
\arguments{
\item{theta_xy}{probability vector for response variables}

\item{P}{Probability matrix}

\item{Q}{Histogram of unique observations}
}
\description{
Calculate the product of theta sums to the power of Qi over Q.
}
\keyword{internal}
