% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsCommon.R
\name{Z_product}
\alias{Z_product}
\title{Z_product}
\usage{
Z_product(theta_z, Q)
}
\arguments{
\item{theta_z}{probability vector}

\item{Q}{Histogram of unique observations}
}
\description{
Calculate the product of theta to power of Qi over Q.
}
\keyword{internal}
