% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsInvalid.R
\name{nps.invalid}
\alias{nps.invalid}
\title{M_Invalid}
\usage{
nps.invalid(Q, l, m, n, N = sum(Q), S = sum(Q))
}
\arguments{
\item{Q}{List of unique observations, should be l\*m\*n length}

\item{l}{|Z|}

\item{m}{|X|}

\item{n}{|Y|}

\item{N}{Number of Repetitions for Nested Sampling}

\item{S}{Number of Starting Points for Nested Sampling}
}
\description{
Calculates the ML_Invalid
}
