% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsMain.R
\name{nps.test}
\alias{nps.test}
\title{Main function of the package.}
\usage{
nps.test(df, l, m, n, N, S)
}
\arguments{
\item{df}{Dataframe with columns z,x and y}

\item{l}{Number of bins used to discretize Z}

\item{m}{Number of bins used to discretize X}

\item{n}{Number of bins used to discretize Y}

\item{N}{Number of Repetitions for Nested Sampling}

\item{S}{Number of Starting Points for Nested Sampling}
}
\value{
result object of the test including the fields: nt, valid, invalid, ratio
}
\description{
Main function of the package.
}
\examples{
nps.test(data.frame(x = runif(3), y = runif(3), z = runif(3)),2,2,2, 3, 3)
}
