% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsMain.R
\name{nps.test}
\alias{nps.test}
\title{Main function of the package.}
\usage{
nps.test(df, l, m, n)
}
\arguments{
\item{df}{Dataframe with columns z,x and y}
}
\value{
result object of the test
}
\description{
Main function of the package.
}
